backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.14.0
pandas>=2.0,<3.0
scikit-learn>=1.4,<2.0
geopy>=2.3.0
python-decouple>=3.8
